rename dump command
